Carousel redirection to marketplace fix
